+++ src/ActivationFunctions/Relu.h
@@ -5,5 +5,5 @@
 class Relu : public ActivationFunction {
     float map(float input) const;
     float getPreValueDerivative(float input) const;
-    float generateRandomWeight(int32_t connectionsIn) const;
+    float generateRandomWeight(size_t connectionsIn) const;
 };
