+++ src/ActivationFunctions/Sigmoid.cpp
@@ -10,7 +10,7 @@ float Sigmoid::getPreValueDerivative(float input) const {
     return input * (1 - input);
 }
 
-float Sigmoid::generateRandomWeight(int32_t connectionsIn) const { // Xavier initialization
+float Sigmoid::generateRandomWeight(size_t connectionsIn) const { // Xavier initialization
     const float r = 1 / std::sqrtf(static_cast<float>(connectionsIn));
 
     std::random_device rd;
